pyproject.toml
[tool.poetry]
name = "replicate-discord-command"
version = "0.1.0"
description = ""
authors = ["Ben Firshman <ben@firshman.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
discord-py-interactions = {git = "https://github.com/interactions-py/library.git", rev = "4.1.1-rc.1"}
httpx = "^0.22.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
